backend/requirements.txt
django
pyodbc
pillow
requests
python-dotenv
django-notifications-hq
liqpay-sdk-python3